+++ src/instructions/ldi.rs
@@ -14,7 +14,7 @@ impl LDI {
     }
 
     pub fn exec(&self, regs: &mut Reg, memory: &Memory) {
-        regs.Rx[self.dr] = memory.read(memory.read((regs.PC + self.pc_offset) as usize) as usize);
+        regs.Rx[self.dr] = memory.read(memory.read(regs.PC + self.pc_offset));
         regs.update_flag(regs.Rx[self.dr]);
     }
 }
